Removed buildinfo
