+++ scripts/ingestion/constants.py
@@ -9,3 +9,6 @@ REGEX = "(\ )+"
 INST_LOC = 'https://popcon.debian.org/by_inst'
 POPCON_CSV = '/data/yellow/vineet/python_files/new_scripts/data_pre_processing/today.csv'
 POPCON_TEXT = '/data/yellow/vineet/python_files/new_scripts/data_pre_processing/today.txt'
+VULNERABILITY_URL = 'https://security-tracker.debian.org/tracker/data/json'
+NIST_URL = 'https://nvd.nist.gov/vuln/detail/'
+VULNERABILITY_DATA = '/data/cyan/guacalytics/raw_data/vulnerability_data/{}/{}/{}'
