+++ scripts/ingestion/database/vulnerability_db_init.py
@@ -1,0 +1,19 @@
+#!/usr/bin/env python
+import sqlite3
+from constants import DB_LOC
+
+conn = sqlite3.connect(DB_LOC)
+cursor = conn.cursor()
+
+def db_init():
+        cursor.execute("""CREATE TABLE IF NOT EXISTS vulnerability_table(
+                source_id integer,
+                package varchar,
+                vulnerability varchar,
+                description varchar,
+                published_date datetime, 
+                last_modified_date datetime,
+                debianbug integer,
+                scope varchar,
+                releases varchar,
+                FOREIGN KEY(source_id) references source_table(source_id) )""")
