+++ scripts/ingestion/parsers/vulnerability_compare.py
@@ -3,11 +3,11 @@ import json
 from bs4 import BeautifulSoup
 from datetime import date
 import sqlite3
+from constants import VULNERABILITY_DATA, NIST_URL, DB_LOC, VULNERABILITY_URL
 
 # Function to fetch data from the URL
 def fetch_data():
-    url = "https://security-tracker.debian.org/tracker/data/json"
-    response = requests.get(url)
+    response = requests.get(VULNERABILITY_URL)
     if response.status_code == 200:
         return response.json()
     else:
@@ -19,14 +19,14 @@ def save_to_json(data):
     year=date.today().year
     mon=date.today().month
     day=date.today().day
-    f='/data/yellow/guacalytics/raw_data/vulnerability_data/{}/{}/{}'.format(year,mon,day)
+    f=VULNERABILITY_DATA.format(year,mon,day)
     with open(f, "w") as file:
         json.dump(data, file)
     file.close()
     return file
 
 def parse_vulnerability_data(vulnerability_id):
-    url = f"https://nvd.nist.gov/vuln/detail/{vulnerability_id}"
+    url = f"{NIST_URL}{vulnerability_id}"
     
     response = requests.get(url)
     if response.status_code == 200:
@@ -56,7 +56,7 @@ def insert_records(cur,name,cve,description,published_date, last_modified_date,d
 
 
 def parser(file):
-    con = sqlite3.connect('/data/yellow/guacalytics/database/bi_multi_tables.db')
+    con = sqlite3.connect(DB_LOC)
     # Check if the file exists locally, if not, download it
     with open (file, 'r') as f:
         parsed_data = json.load(f)
