+++ scripts/ingestion/parsers/vulnerability_parser.py
@@ -2,12 +2,13 @@
 import requests
 import sqlite3
 from vulnerability_main import insert_records
+from constants import DB_LOC, VULNERABILITY_DATA
 
 # Define the main parser function
 def parser():
-    con = sqlite3.connect('/data/yellow/guacalytics/database/bi_multi_tables.db')
+    con = sqlite3.connect(DB_LOC)
     # Check if the file exists locally, if not, download it
-    response = requests.get("https://security-tracker.debian.org/tracker/data/json")
+    response = requests.get(VULNERABILITY_DATA)
 
         # Parse the data as JSON
     parsed_data = response.json()
@@ -32,6 +33,7 @@ def parser():
             # to insert records into the table
             insert_records(cur,name,cve,description,debianbug,scope,releases) 
             con.commit()
+    con.close()
 
 if __name__ == "__main__":
     parser()
