+++ scripts/vulnerability_main.py
@@ -1,11 +1,13 @@
 #!/usr/bin/env python
 import requests
 from bs4 import BeautifulSoup
+from database.vulnerability_db_init import db_init,conn
+from constants import NIST_URL
 
 INSERT_RECS = '''INSERT OR IGNORE INTO vulnerability_table (source_id,package,vulnerability,description,published_date, last_modified_date,debianbug,scope,releases) VALUES (?,?,?,?,?,?,?,?,?)'''
 
 def parse_vulnerability_data(vulnerability_id):
-    url = f"https://nvd.nist.gov/vuln/detail/{vulnerability_id}"
+    url = f"{NIST_URL}{vulnerability_id}"
     
     response = requests.get(url)
     if response.status_code == 200:
@@ -24,17 +26,7 @@ def parse_vulnerability_data(vulnerability_id):
     return None, None
 
 def insert_records(cur,name,cve,description,debianbug,scope,releases):
-    cur.execute("""CREATE TABLE IF NOT EXISTS vulnerability_table(
-                source_id integer,
-                package varchar,
-                vulnerability varchar,
-                description varchar,
-                published_date datetime, 
-                last_modified_date datetime,
-                debianbug integer,
-                scope varchar,
-                releases varchar,
-                FOREIGN KEY(source_id) references source_table(source_id) )""") 
+    db_init()
     cur.execute('''select source_id,source_name from source_table where source_name ='{}' limit 1'''.format(name))
     id = cur.fetchone()
     published_date, last_modified_date = parse_vulnerability_data(cve)
